Strange windows permissions when generating html-report
The build now happens on a windows runner, to be able to support windows. However when writing a file, the binary is not setting the correct permissions, the file cannot be read. If this can’t be resolved, I will need to split the build into a windows and linux run.

Signed-off-by: quobix <dave@quobix.com>