Upgraded to libopenapi v0.13
A very different way of handling indexes and remote and local filesystems. The addition of a new `-r` or `—remote` flag for all commands will turn on automatic handling of remote referneces (be them local or remote) and will attempt to resolve them.

This overrides the `—base` / `-p` flag if set.

Signed-off-by: quobix <dave@quobix.com>