At limit commit read file content at previous commit
When processing the earliest commit of the commit history, no data from the file
outside of the history is present. This means that no changes can be calculated
for the first commit.

This issue was likely always present but surfaced by the introduction of the
`--global-revisions` flag. Previously, the issue would only appear when reaching
the commit that introduced a file.

To address the problem, we now read the file's content at the previous commit in the
repo.

This still does not resolve the problem caused by a file simply not being
present at the previous commit.