wrapped git command to prevent shell issues
Trying to debug https://github.com/pb33f/openapi-changes/issues/201

The date parsing is handled by a “parse anything’ lib, which is great, but also confusing as there should never be an invalid date. When trying to re-run this command locally.  The ‘||’ delimiter was being interpreted as ‘or’ instead of a format string. I changed the delimiter and upgraded it to a quoted string.

Lets see if this solves it.